requirements.txt
# Img2CAD: Complete Project Dependencies
# This file contains all dependencies for both LlamaFT and TrAssembler modules

# Core Python libraries
torch==2.5.1+cu118
torchvision==0.20.1+cu118
numpy==1.26.2
opencv-python==4.8.1.78
h5py==3.10.0
tqdm==4.67.1
pillow==11.0.0
scipy==1.11.4
trimesh==4.0.9
matplotlib==3.8.2

# Deep learning frameworks and training
pytorch-lightning==2.5.1
transformers==4.46.1
accelerate==1.1.1

# Vision-language models and embeddings
qwen-vl-utils==0.0.8

# Transformer utilities and attention mechanisms
einops==0.7.0
torch-scatter==2.1.2

# Configuration management
hydra-core==1.3.2
omegaconf==2.3.0

# Fine-tuning and PEFT (for LlamaFT)
trl==0.11.4
peft==0.12.0
datasets==3.1.0
bitsandbytes==0.44.1

# Hugging Face ecosystem
huggingface-hub==0.34.4

# Logging and monitoring
wandb==0.17.0
tensorboard==2.15.1

# Visualization and UI
visdom==0.2.4

# CUDA rendering (optional, for GPU acceleration)
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@729261dc64c4241ea36efda84fbf532cc8b425b8